+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.2.0
+
+- removed Bot.RunForever(), and now Bot.Run() doesn't require a channel as argument, added Bot.Quit()
+  to stop processing events
+
 ## v0.1.0
 
 - initial release